refactor(AdminPage): Exploded file into manageable component files
